fix resource compute + test structure
