+++ shipcat_cli/Cargo.toml
@@ -21,8 +21,8 @@ name = "shipcat"
 path = "src/main.rs"
 
 [dependencies]
-clap = "2.31.2"
 shipcat_definitions = { path = "../shipcat_definitions" }
+clap = "2.31.2"
 error-chain = "0.12.0"
 log = "0.4.5"
 loggerv = "0.7.1"
@@ -36,15 +36,13 @@ serde_yaml = "0.8.5"
 slack-hook = "0.8.0"
 jenkins_api = "0.5.1"
 #tera = "0.11.16"
-#walkdir = "2.1.4"
 threadpool = "1.7.1"
 chrono = { version = "0.4.6", features = ["serde"] }
-#semver = { version = "0.9.0", features = ["serde"] }
+semver = { version = "0.9.0", features = ["serde"] }
 base64 = "0.9.3"
 dirs = "1.0.3"
 libc = "0.2.43"
 walkdir = "2.2.5"
-semver = "0.9.0"
 
 [dependencies.petgraph]
 features = ["serde-1"]
