+++ shipcat_cli/src/get.rs
@@ -1,5 +1,4 @@
 /// This file contains the `shipcat get` subcommand
-use shipcat_definitions::math::ResourceTotals;
 use std::collections::BTreeMap;
 
 use super::{Result, Manifest, Config};
@@ -141,107 +140,31 @@ pub fn apistatus(conf: &Config, region: &str) -> Result<()> {
 }
 
 // ----------------------------------------------------------------------------
-/*
-use super::structs::Resources;
-use shipcat_definitions::manifest::math::ResourceTotals;
 
-#[derive(Serialize, Default)]
-pub struct ResourceBreakdown {
-    /// Total totals
-    pub totals: ResourceTotals,
-    /// A partition of totals info teams
-    pub teams: BTreeMap<String, ResourceTotals>,
-}
+use shipcat_definitions::ResourceBreakdown;
 
-
+/// Resource use for a single region
 pub fn resources(conf: &Config, region: &str) -> Result<()> {
-    let services = Manifest::available()?;
-    let mut bd = ResourceBreakdown::default(); // zero for all the things
-
-    let mut sum : Resources<f64> = Resources::default();
-    let mut extra : Resources<f64> = Resources::default(); // autoscaling limits
-
-    for svc in services {
-        let mf = Manifest::basic(&svc, &conf, None)?;
-        if mf.external || !mf.regions.contains(&region.to_string()) {
-            continue; // only care about kube services in the current region
-        }
-        if let Some(ref md) = mf.metadata {
-            let ResourceTotals { base: sb, extra: se } = mf.compute_resource_totals()?;
-            sum += sb.clone();
-            extra += se.clone();
-            bd.teams.entry(md.team.clone())
-                .and_modify(|e| {
-                    e.base += sb.clone();
-                    e.extra += se.clone();
-                })
-                .or_insert_with(|| {
-                    ResourceTotals { base: sb, extra: se }
-                }
-            );
-        } else {
-            bail!("{} service does not have resources specification and metadata", mf.name)
-        }
-    }
-    // convert gigs for all teams
-    for (_, mut teamtot) in &mut bd.teams {
-        teamtot.base.round();
-        teamtot.extra.round();
-    }
-    // overall totals:
-    sum.round();
-    extra.round();
-    bd.totals.base = sum;
-    bd.totals.extra = extra;
-
-    print!("{}\n", serde_yaml::to_string(&bd)?);
+    let bd = Manifest::resources(conf, region)?.normalise();
+    print!("{}\n", serde_json::to_string_pretty(&bd)?);
     Ok(())
 }
 
+/// Resources for all regions
 pub fn totalresources(conf: &Config) -> Result<()> {
-    let services = Manifest::available()?;
-    let mut bd = ResourceBreakdown::default(); // zero for all the things
-
-    let mut sum : Resources<f64> = Resources::default();
-    let mut extra : Resources<f64> = Resources::default(); // autoscaling limits
-
-    for svc in services {
-        let tmpmf = Manifest::basic(&svc, &conf, None)?;
-        if tmpmf.external {
-            continue; // only care about kube services
-        }
-        for region in tmpmf.regions {
-            let mf = Manifest::mocked(&svc, &conf, &region)?;
-            if let Some(ref md) = mf.metadata {
-                let ResourceTotals { base: sb, extra: se } = mf.compute_resource_totals()?;
-                sum += sb.clone();
-                extra += se.clone();
-                bd.teams.entry(md.team.clone())
-                    .and_modify(|e| {
-                        e.base += sb.clone();
-                        e.extra += se.clone();
-                    })
-                    .or_insert_with(|| {
-                        ResourceTotals { base: sb, extra: se }
-                    }
-                );
-            } else {
-                bail!("{} service does not have resources specification and metadata", mf.name)
-            }
+    let mut bd = ResourceBreakdown::new(conf.teams.clone()); // zero for all the things
+    for r in conf.regions.keys() {
+        let res = Manifest::resources(conf, r)?;
+        bd.totals.base += res.totals.base;
+        bd.totals.extra += res.totals.extra;
+        for t in &conf.teams {
+            let rhs = &res.teams[&t.name];
+            let e = bd.teams.get_mut(&t.name).unwrap(); // exists by ResourceBreakdown::new
+            e.base += rhs.base.clone();
+            e.extra += rhs.extra.clone();
         }
     }
-    // convert gigs for all teams
-    for (_, mut teamtot) in &mut bd.teams {
-        teamtot.base.round();
-        teamtot.extra.round();
-    }
-    // overall totals:
-    sum.round();
-    extra.round();
-    bd.totals.base = sum;
-    bd.totals.extra = extra;
-
-    print!("{}\n", serde_yaml::to_string(&bd)?);
+    bd = bd.normalise();
+    print!("{}\n", serde_json::to_string_pretty(&bd)?);
     Ok(())
 }
-*/
