+++ shipcat_definitions/src/filebacked.rs
@@ -233,50 +233,3 @@ impl Manifest {
     }
 
 }
-
-#[cfg(test)]
-mod tests {
-    use tests::setup;
-    use super::Config;
-    use super::Manifest;
-
-    #[test]
-    fn manifest_test() {
-        setup();
-        let conf = Config::read().unwrap();
-        let mf = Manifest::completed("fake-storage", &conf, "dev-uk".into()).unwrap();
-        // verify datahandling implicits
-        let dh = mf.dataHandling.unwrap();
-        let s3 = dh.stores[0].clone();
-        assert!(s3.encrypted.unwrap());
-        assert_eq!(s3.fields[0].encrypted.unwrap(), false); // overridden
-        assert_eq!(s3.fields[1].encrypted.unwrap(), true); // cascaded
-        assert_eq!(s3.fields[0].keyRotator, None); // not set either place
-        assert_eq!(s3.fields[1].keyRotator, Some("2w".into())); // field value
-    }
-
-    #[test]
-    fn templating_test() {
-        setup();
-        let conf = Config::read().unwrap();
-        let mf = Manifest::completed("fake-ask", &conf, "dev-uk".into()).unwrap();
-
-        // verify templating
-        let env = mf.env;
-        assert_eq!(env["CORE_URL"], "https://woot.com/somesvc".to_string());
-        // check values from Config - one plain, one as_secret
-        assert_eq!(env["CLIENT_ID"], "FAKEASKID".to_string());
-        assert!(env.get("CLIENT_SECRET").is_none()); // moved to secret
-        let sec = mf.secrets;
-        assert_eq!(sec["CLIENT_SECRET"], "FAKEASKSECRET".to_string()); // via reg.kong consumers
-        assert_eq!(sec["FAKE_SECRET"], "hello".to_string()); // NB: ACTUALLY IN_VAULT
-
-        let configs = mf.configs.clone().unwrap();
-        let configini = configs.files[0].clone();
-        let cfgtpl = configini.value.unwrap();
-        print!("{:?}", cfgtpl);
-        assert!(cfgtpl.contains("CORE=https://woot.com/somesvc"));
-        assert!(cfgtpl.contains("CLIENT_ID"));
-        assert!(cfgtpl.contains("CLIENT_ID=FAKEASKID"));
-    }
-}
