+++ shipcat_definitions/src/reducers.rs
@@ -1,3 +1,4 @@
+use config::Team;
 use semver::Version;
 use std::collections::BTreeMap;
 use super::{Result, Config, Manifest};
@@ -9,7 +10,6 @@ impl Manifest {
     ///
     /// Services without a hardcoded version are not returned.
     pub fn get_versions(conf: &Config, region: &str) -> Result<BTreeMap<String, Version>> {
-        use semver::Version;
         let services = Manifest::available()?;
         let mut output = BTreeMap::new();
 
@@ -68,46 +68,72 @@ impl Manifest {
         }
         Ok(output)
     }
-
 }
 
+/// Complete breakdown of resource usage in total, and split by team.
+///
+/// Normally this is computed by `Manifest::resources` for a region-wide total.
+/// Looping over all regions is possible in the CLI.
+#[derive(Serialize)]
+pub struct ResourceBreakdown {
+    /// Total totals
+    pub totals: ResourceTotals,
+    /// A partition of totals info teams
+    pub teams: BTreeMap<String, ResourceTotals>,
+}
 
-
-#[cfg(test)]
-mod tests {
-    use tests::setup;
-    use super::Config;
-    use super::Manifest;
-    use semver::Version;
-
-    #[test]
-    fn get_versions() {
-        setup();
-        let conf = Config::read().unwrap();
-        let vers = Manifest::get_versions(&conf, "dev-uk").unwrap();
-
-        assert_eq!(vers.len(), 1); // only one of the services has a version
-        assert_eq!(vers["fake-ask"], Version::new(1, 6, 0));
+impl ResourceBreakdown {
+    /// Constructor to ensure all valid teams are filled in
+    pub fn new(tx: Vec<Team>) -> ResourceBreakdown {
+        let mut teams = BTreeMap::new();
+        for t in tx {
+            teams.insert(t.name, ResourceTotals::default());
+        }
+        ResourceBreakdown { teams, totals: ResourceTotals::default() }
     }
 
-    #[test]
-    fn get_images() {
-        setup();
-        let conf = Config::read().unwrap();
-        let vers = Manifest::get_images(&conf, "dev-uk").unwrap();
-
-        assert_eq!(vers.len(), 2); // every service gets an image
-        assert_eq!(vers["fake-ask"], "quay.io/babylonhealth/fake-ask");
-        assert_eq!(vers["fake-storage"], "nginx");
+    /// Round all numbers to gigs and full cores (for all teams)
+    pub fn normalise(mut self) -> Self {
+        for (_, mut tt) in &mut self.teams {
+            tt.base.round();
+            tt.extra.round();
+        }
+        self.totals.base.round();
+        self.totals.extra.round();
+        self
     }
+}
 
-    #[test]
-    fn get_codeowners() {
-        setup();
-        let conf = Config::read().unwrap();
-        let cos = Manifest::get_codeowners(&conf, "dev-uk").unwrap();
+use super::structs::Resources;
+use super::math::ResourceTotals;
 
-        assert_eq!(cos.len(), 1); // services without owners get no listing
-        assert_eq!(cos[0], "services/fake-ask/* @clux");
+impl Manifest {
+    /// Compute resource usage for all available manifests in a region.
+    pub fn resources(conf: &Config, region: &str) -> Result<ResourceBreakdown> {
+        let services = Manifest::available()?;
+        let mut bd = ResourceBreakdown::new(conf.teams.clone()); // zero for all the things
+
+        let mut sum : Resources<f64> = Default::default();
+        let mut extra : Resources<f64> = Default::default(); // autoscaling limits
+
+        for svc in services {
+            let mf = Manifest::basic(&svc, &conf, None)?;
+            if mf.external || !mf.regions.contains(&region.to_string()) {
+                continue; // only care about kube services in the current region
+            }
+            if let Some(ref md) = mf.metadata {
+                let ResourceTotals { base: sb, extra: se } = mf.compute_resource_totals()?;
+                sum += sb.clone();
+                extra += se.clone();
+                let e = bd.teams.get_mut(&md.team).unwrap(); // exists by ResourceBreakdown::new
+                e.base += sb.clone();
+                e.extra += se.clone();
+            } else {
+                bail!("{} service does not have resources specification and metadata", mf.name)
+            }
+        }
+        bd.totals.base = sum;
+        bd.totals.extra = extra;
+        Ok(bd)
     }
 }
